rename: web과 console 패키지 분리
