+++ src/main/java/chess/console/Application.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.console;
 
 import chess.controller.ChessController;
